requirements.txt
playwright~=1.49.1
pytest~=8.3.4
behave~=1.2.6
allure-python-commons~=2.13.5
asyncio~=3.4.3
